flask_api/requirements.txt
Flask
psycopg2-binary
python-dotenv
pandas